# Use mimalloc as default allocator