From https://github.com/microsoft/mimalloc:
> In our benchmarks (see [below](https://github.com/microsoft/mimalloc#performance)), mimalloc outperforms other leading allocators (jemalloc, tcmalloc, Hoard, etc), and often uses less memory. A nice property is that it does consistently well over a wide range of benchmarks. There is also good huge OS page support for larger server programs.


Release Notes:

- Changed default allocator to mimalloc.