@osiewicz 2024-01-31
Hey, thanks for the PR! How significant is the performance improvement? I believe we have a script called "measure" (or something like that) to measure frame time.