@TENX-S 2024-01-31
For frame time:
![perf](https://github.com/zed-industries/zed/assets/40336192/a415ae41-8be1-4843-b700-c49ef8b8b14f)

For memory usage:
![WeChat851680969783f9d155e827538540d339](https://github.com/zed-industries/zed/assets/40336192/01c7dc52-311b-4406-9410-cd8d41fb447e)