@TENX-S 2024-01-31
This is how I benchmark frame time:
1. restart computer
2. close all unrelated processes
2. disable language server in settings
3. follow the instructions from #4191
4. open zed project and open `crates/editor/src/editor.rs`
5. `gg` go to the top of `editor.rs`
6. press `Shift` and `]`(a vim motion for jumping to next paragraph), hold it until at bottom. See video below for example.
Repeat 1~6 for Zed with mimalloc allocator again, we get the following result:
![perf](https://github.com/zed-industries/zed/assets/40336192/4b80d3b1-7a63-4548-8fcf-031605a1c941)

https://github.com/zed-industries/zed/assets/40336192/9a2699cf-0f29-4868-948b-e90d2b2c7ded

For memory usage, open golang source code for example:
![WeChat851680969783f9d155e827538540d339](https://github.com/zed-industries/zed/assets/40336192/84bf0d43-c903-4efa-b14b-628747da6462)