@Titaniumtown 2024-01-31
Just a note, since #5396 exists, using mimalloc on the web is complicated in my experience, so the usage of mimalloc should be limited to only native binaries.