@as-cii 2024-01-31
That's a pretty impressive performance improvement! Thanks @TENX-S!

As for web support, I think @Titaniumtown has a good point but I also think compiling Zed for the web is gonna involve porting other parts of the app that currently assume we're running on macOS, so I think we can deal with it when we cross that bridge.