@mrnugget 2024-02-01
@TENX-S very cool! Out of curiosity: you did compare Zed builds in `--release` mode, right?