@michealp-coder 2024-03-25
@as-cii 

I see this was merged, but it's unclear to me if this has landed in Preview/Release yet.

If it has been committed, any reason to keep this 'open'?

(Apologies for my lack of understanding. Btw, love Zed. Please start charging for it. I'm a solo developer and want to support your all. Even though I won't benefit from the collaboration capabilities, I would like to pay for the core editor via donation or pricing)