@michealp-coder 2024-03-26
I'm not the most well versed in how Github indicates things. It's probably
my misunderstanding, apologies.

On Tue, Mar 26, 2024 at 12:38 AM Thorsten Ball ***@***.***>
wrote:

> @michealp-coder <https://github.com/michealp-coder> yep, it's been merged
> and released! The PR is closed — what do you mean is 'open'?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/zed-industries/zed/pull/7140#issuecomment-2019433635>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AUEN7J7PV4RW5TXJ3TU5YT3Y2D3V7AVCNFSM6AAAAABCSU7LVGVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDAMJZGQZTGNRTGU>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>