@michealp-coder 2024-03-27
Was https://zed.dev updated with the new Memory Footprint usage after this change to mimalloc?

I see on https://zed.dev `257mb` of memory is consumed by Zed.

Based on the numbers reported in this PR and my own usage of Zed, it's considerably less than 257mb.