@mrnugget 2024-03-27
Fresh instance of `Zed.app`: 
![screenshot-2024-03-27-06 34 04@2x](https://github.com/zed-industries/zed/assets/1185253/9de2b15a-b1bf-478f-824f-4c1b927ee51e)

I still think it's ~accurate. But that being said we don't regularly update the stats on that page.