@mrnugget 2024-04-16
Coming back to this after seeing `mimalloc` being really aggressive with memory retention and allocation, triggered by https://github.com/zed-industries/zed/issues/10525#issuecomment-2058458892 and https://github.com/zed-industries/zed/pull/10321#issuecomment-2056763742.

I ran some benchmarks now, on `main`, on my M3 Max.

### Methodology:

- Run `cargo run --release <projectfolder>`
- Close every open file
- Only keep project panel open
- Quit process
- Run `cargo run --release <projectfolder>` again
- Open `Activity Monitor.app`
- Have both — `Zed` and `Activity Monitor` — visible
- Write down `Before loading`
- Open file
- Wait 5sec, write down `After loading`
- Close file
- Wait 5sec, write down `After closing`


### Results:

| What                          | Allocator  | Before loading | After loading | After closing | Comment                            |
| ----------------------------- | ---------- | -------------- | ------------- | ------------- | ---------------------------------- |
| 132MB `.tex` file             | `mimalloc` | 200 MB         | 767 MB        | 768 MB        | No tex extension installed         |
| 10KB `.zshrc` file            | `mimalloc` | 169 MB         | 175 MB        | 182 MB        | -                                  |
| 403KB `editor.rs` file in Zed | `mimalloc` | 167 MB         | 502 MB        | 208 MB        | Waited until rust-analyzer started |
| 132MB `.tex` file             | `default`  | 116 MB         | 675 MB        | 457 MB        | No tex extension installed         |
| 10KB `.zshrc` file            | `default`  | 150 MB         | 156 MB        | 156 MB        | -                                  |
| 403KB `editor.rs` file in Zed | `default`  | 156 MB         | 477 MB        | 173 MB        | Waited until rust-analyzer started |

### Frametimes:

(Following these instructions: https://github.com/zed-industries/zed/pull/7140#issuecomment-1918887685 Except that I didn't restart and didn't close all other processes, I have 128GB of memory)

![screenshot-2024-04-16-15 16 58@2x](https://github.com/zed-industries/zed/assets/1185253/190e7489-5baf-40a0-96b4-1d04f3c65523)

Note the `x` axis here.

### Conclusion?

It looks to me like `mimalloc` isn't a clear winner. While there might be some improvements on frametime, it's not that clear cut. 

At the same time, `mimalloc` seems to retain more memory and allocate larger pages more aggressively.

I think we can ignore the differences in `Before loading` on each test run, since that depends on window size, whether I just moved it, etc. But I wanted "realistic" circumstances, since that's what causes users to open tickets: they open Zed, note memory usage, open ticket, they don't check again on a clean machine.)

From my perspective, it seems like we can revert the `mimalloc-as-default` change and Zed would use less memory, free up memory more easily, and only maaaybe take a slight performance hit.

Thoughts?