@mocenigo 2024-04-16
> Oh I need to redo this! The results vary wildly depending on whether Activity Monitor or Zed is focused

I suppose that was with the macOS standard hoard-derived allocator? I know that one can recompile MiMalloc to have more aggressive releasing, and/or to accept that some threads may share memory arenas (I would not do the latter), but maybe this is too much. Anyway, after closing the 132Mb file in Zed, on my machine the memory goes down to 160Mb, but I am of course also switching my focus to Activity monitor.

Then, if I repeat the process (with both allocators), each time some more memory is added, so there may still be some mem leak.

Finally, even with the macOS standard allocator, if I start the LaTeX extension, the latter starts to consume a lot of memory, but if I close the file, the memory still goes on (the extension is not terminated) and if I go and remove the extension, memory allocation is stopped but the memory is not released back to the system (or at least not all of it).